more work on file load and save.

+++ applications/wxDemosSandbox/sdkDemos/demos/MultiRayCasting.cpp
@@ -95,8 +95,8 @@ class LineOfSightRayCastEntity: public DemoEntity
 		glBegin(GL_LINES);
 		for (dCustomControllerManager<dRayCastRecord>::dListNode* node = m_casterManager->GetFirst(); node; node = node->GetNext()) {
 			dRayCastRecord* const ray = &node->GetInfo();
-			glVertex3f(ray->m_p0.m_x, ray->m_p0.m_y, ray->m_p0.m_z);
-			glVertex3f(ray->m_p1.m_x, ray->m_p1.m_y, ray->m_p1.m_z);
+			glVertex3f(GLfloat(ray->m_p0.m_x), GLfloat(ray->m_p0.m_y), GLfloat(ray->m_p0.m_z));
+			glVertex3f(GLfloat(ray->m_p1.m_x), GLfloat(ray->m_p1.m_y), GLfloat(ray->m_p1.m_z));
 		}
 		glEnd();
 
@@ -106,7 +106,7 @@ class LineOfSightRayCastEntity: public DemoEntity
 		glBegin(GL_POINTS);
 		for (dCustomControllerManager<dRayCastRecord>::dListNode* node = m_casterManager->GetFirst(); node; node = node->GetNext()) {
 			dRayCastRecord* const ray = &node->GetInfo();
-			glVertex3f(ray->m_p1.m_x, ray->m_p1.m_y, ray->m_p1.m_z);
+			glVertex3f(GLfloat(ray->m_p1.m_x), GLfloat(ray->m_p1.m_y), GLfloat(ray->m_p1.m_z));
 		}
 		glEnd();
 		glPointSize(1.0f);
